提交闪光算法

+++ api/Preferred.Api/BusinessModule/ZwavAnalysis/Model/Table/Tb_ZwavFile.cs
@@ -6,7 +6,7 @@ namespace Preferred.Api.Models
     {
         public int Id { get; set; }
         public string OriginalName { get; set; }
-        public long FileSize { get; set; }        // 你表里 INT；建议后面改 BIGINT
+        public int FileSize { get; set; }        // 你表里 INT；建议后面改 BIGINT
         public string StoragePath { get; set; }
          public string ExtractPath { get; set; }
